requirements/common.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/common.in
#
aiodns==2.0.0             # via aiohttp
aiohttp[speedups]==3.6.2
aiopg==1.0.0
aioredis==1.3.0
async-timeout==3.0.1      # via aiohttp, aioredis
attrs==19.3.0             # via aiohttp
brotlipy==0.7.0           # via aiohttp
cachetools==3.1.1         # via google-auth
cchardet==2.1.5           # via aiohttp
certifi==2019.9.11        # via requests, sentry-sdk
cffi==1.13.2              # via brotlipy, pycares
chardet==3.0.4            # via aiohttp, requests
envparse==0.2.0
google-api-core[grpc]==1.14.3  # via google-cloud-core, google-cloud-logging, google-cloud-pubsub
google-auth==1.7.1        # via google-api-core
google-cloud-core==1.0.3  # via google-cloud-logging
google-cloud-logging==1.14.0
google-cloud-pubsub==1.0.2
googleapis-common-protos[grpc]==1.6.0  # via google-api-core, grpc-google-iam-v1
grpc-google-iam-v1==0.12.3  # via google-cloud-pubsub
grpcio==1.25.0            # via google-api-core, googleapis-common-protos, grpc-google-iam-v1
gunicorn==20.0.0
hiredis==1.0.1            # via aioredis
idna==2.8                 # via requests, yarl
multidict==4.5.2          # via aiohttp, yarl
protobuf==3.10.0          # via google-api-core, googleapis-common-protos
psycopg2-binary==2.8.4    # via aiopg
pyasn1-modules==0.2.7     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycares==3.0.0            # via aiodns
pycparser==2.19           # via cffi
pytz==2019.3              # via google-api-core
pyyaml==5.1.2
requests==2.22.0          # via google-api-core
rsa==4.0                  # via google-auth
sentry-sdk==0.13.2
six==1.13.0               # via google-api-core, google-auth, grpcio, protobuf
urllib3==1.25.7           # via requests, sentry-sdk
yarl==1.3.0               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.2        # via google-api-core, google-auth, gunicorn, protobuf
